* upstream/master: (40 commits)
  Fix CI job naming
  [ML] disallow autoscaling downscaling in two trained model assignment scenarios (#88623)
  Add "Vector Search" area to changelog schema
  [DOCS] Update API key API (#88499)
  Enable the pipeline on the feature branch (#88672)
  Adding the ability to register a PeerFinderListener to Coordinator (#88626)
  [DOCS] Fix transform painless example syntax (#88364)
  [ML] Muting InternalCategorizationAggregationTests testReduceRandom (#88685)
  Fix double rounding errors for disk usage (#88683)
  Replace health request with a state observer. (#88641)
  [ML] Fail model deployment if all allocations cannot be provided (#88656)
  Upgrade to OpenJDK 18.0.2+9 (#88675)
  [ML] make bucket_correlation aggregation generally available (#88655)
  Adding cardinality support for random_sampler agg (#86838)
  Use custom task instead of generic AckedClusterStateUpdateTask (#88643)
  Reinstate test cluster throttling behavior (#88664)
  Mute testReadBlobWithPrematureConnectionClose
  Simplify plugin descriptor tests (#88659)
  Add CI job for testing more job parallelism
  [ML] make deployment infer requests fully cancellable (#88649)
  ...

+++ server/src/main/java/org/elasticsearch/cluster/DiskUsage.java
@@ -97,7 +97,7 @@ public class DiskUsage implements ToXContentFragment, Writeable {
         if (totalBytes == 0) {
             return 100.0;
         }
-        return 100.0 * ((double) freeBytes / totalBytes);
+        return 100.0 * freeBytes / totalBytes;
     }
 
     public double getUsedDiskAsPercentage() {
